Relocate plugin-based commands to 'pc-pluginManager'
