+++ src/Powercord/plugins/pc-pluginManager/commands/enable.js
@@ -1,22 +1,22 @@
 module.exports = {
-  command: 'disable',
-  description: 'Allows you to disable and unload a plugin from the given list.',
-  usage: '{c} [pluginID]',
+  command: 'enable',
+  description: 'Allows you to re-/enable a selected plugin from the given list.',
+  usage: '{c} [ plugin ID ]',
 
   func (args) {
     let result;
 
     if (powercord.pluginManager.plugins.has(args[0])) {
-      if (!powercord.pluginManager.isEnabled(args[0])) {
-        result = `->> ERROR: Tried to unload a non-loaded plugin
+      if (powercord.pluginManager.isEnabled(args[0])) {
+        result = `->> ERROR: Tried to load an already loaded plugin!
             (${args[0]})`;
       } else {
-        result = `+>> SUCCESS: Plugin unloaded
-              (${args[0]})`;
-        powercord.pluginManager.disable(args[0]);
+        powercord.pluginManager.enable(args[0]);
+        result = `+>> SUCCESS: Plugin loaded!
+            (${args[0]})`;
       }
     } else {
-      result = `->> ERROR: Tried to disable a non-installed plugin
+      result = `->> ERROR: Tried to enable a non-installed plugin!
           (${args[0]})`;
     }
 
@@ -28,7 +28,7 @@ module.exports = {
 
   autocompleteFunc (args) {
     const plugins = powercord.pluginManager.getPlugins()
-      .sort((a, b) => a < b ? -1 : 1 || 0)
+      .sort((a, b) => a - b)
       .map(plugin => powercord.pluginManager.plugins.get(plugin));
 
     if (args.length > 1) {
@@ -37,8 +37,7 @@ module.exports = {
 
     return {
       commands: plugins
-        .filter(plugin => plugin.pluginID !== 'pc-commands' &&
-          plugin.pluginID.includes(args[0].toLowerCase()))
+        .filter(plugin => plugin.pluginID.includes(args[0].toLowerCase()))
         .map(plugin => ({
           command: plugin.pluginID,
           description: plugin.manifest.description
