+++ src/Powercord/plugins/pc-pluginManager/commands/index.js
@@ -1,0 +1,7 @@
+require('fs')
+  .readdirSync(__dirname)
+  .filter(file => file !== 'index.js')
+  .forEach(filename => {
+    const moduleName = filename.split('.')[0];
+    exports[moduleName] = require(`${__dirname}/${filename}`);
+  });
