+++ src/Powercord/plugins/pc-pluginManager/commands/plugins.js
@@ -1,4 +1,3 @@
-
 module.exports = {
   command: 'plugins',
   aliases: [ 'plist' ],
